+++ packages/dashboard/src/components/DrivesTableHead.js
@@ -10,6 +10,7 @@ import AddCircleOutlineIcon from '@material-ui/icons/AddCircleOutline'
 import { makeStyles } from '@material-ui/core'
 
 const headCells = [
+  { id: 'newItem', numeric: false },
   { id: 'seedingStatus', numeric: false, label: 'Status', width: '40px' },
   {
     id: 'key',