+++ packages/sdk/src/services/keys-updater.service.js
@@ -19,10 +19,10 @@ module.exports = {
 
   events: {
     async 'keys.created' (ctx) {
-      await ctx.call('seeder.seed', { keys: ctx.params.keys.map(({ key }) => key) })
+      await ctx.call('seeder.seed', { keys: ctx.params.keys.map(({ key }) => key), created: true })
     },
     async 'keys.updated' (ctx) {
-      await ctx.call('seeder.seed', { keys: ctx.params.keys.map(({ key }) => key) })
+      await ctx.call('seeder.seed', { keys: ctx.params.keys.map(({ key }) => key), created: false })
     }
   },