+++ packages/sdk/src/services/metrics.service.js
@@ -1,14 +1,15 @@
 const { resolve } = require('path')
+const { promisify } = require('util')
 
 const { MetricsDatabase } = require('@geut/permanent-seeder-database')
 const top = require('process-top')()
-const trammel = require('trammel')
+const { fsize } = require('nodejs-fs-utils')
 const pMemoize = require('p-memoize')
 const isOnline = require('is-online')
 
 const { Config } = require('../mixins/config.mixin')
 
-const dirSize = pMemoize(trammel, { maxAge: 2000 })
+const dirSize = pMemoize(promisify(fsize), { maxAge: 2000 })
 
 module.exports = {
   name: 'metrics',
@@ -22,11 +23,10 @@ module.exports = {
 
   events: {
     'seeder.drive.*': {
-      throttle: 200,
       async handler (ctx) {
         const timestamp = Date.now()
-        const { eventName: event, params: { key } } = ctx
-        await this.saveStats({ key, timestamp, event })
+        const { eventName: event, params: { key, ...rest } } = ctx
+        await this.saveStats({ key, timestamp, event, ...rest })
       }
     },
     'seeder.networker.peer.*': {
@@ -74,7 +74,7 @@ module.exports = {
         let disk = ''
 
         try {
-          disk = await dirSize(this.config.path, { stopOnError: true })
+          disk = await dirSize(this.config.path, { skipErrors: true })
         } catch (error) {
           console.error(error)
         }
@@ -108,8 +108,10 @@ module.exports = {
           return
         }
 
-        data.host = await this.getHostStats()
-        if (data.event !== 'seeder.drive.remove') {
+        if (data.event === 'seeder.drive.download-started' || data.event === 'seeder.drive.download-finished') {
+          data.host = await this.getHostStats()
+        }
+        if (data.event !== 'seeder.drive.download' && data.event !== 'seeder.drive.remove') {
           data.peers = await this.broker.call('seeder.drivePeers', { key: data.key })
         }
         return this.database.add(data)